+++ Sources/EudiWalletKit/Services/BlePresentationService.swift
@@ -25,10 +25,8 @@ import MdocDataTransfer18013
 public final class BlePresentationService: @unchecked Sendable, PresentationService {
 	var bleServerTransfer: MdocGattServer
 	public var status: TransferStatus = .initializing
-	var continuationQrCode: CheckedContinuation<String, Error>?
 	var continuationRequest: CheckedContinuation<UserRequestInfo, Error>?
-	var continuationResponse: CheckedContinuation<Void, Error>?
-	var handleSelected: ((Bool, RequestItems?) -> Void)?
+	var handleSelected: ((Bool, RequestItems?) async -> Void)?
 	var deviceEngagement: String?
 	var request: UserRequestInfo?
 	public var flow: FlowType { .ble }
@@ -42,11 +40,17 @@ public final class BlePresentationService: @unchecked Sendable, PresentationServ
 
 	/// The holder app should present the returned code to the verifier
 	/// - Returns: The image data for the QR code
-	public func startQrEngagement() async throws -> String? {
-		return try await withCheckedThrowingContinuation { c in
-			continuationQrCode = c
-			self.bleServerTransfer.performDeviceEngagement()
+	public func startQrEngagement(secureAreaName: String?, crv: CoseEcCurve) async throws -> String {
+		if bleServerTransfer.unlockData == nil {
+			var unlockData = [String: Data]()
+			for (id, key) in bleServerTransfer.devicePrivateKeys {
+				let ud = try await key.secureArea.unlockKey(id: id)
+				if let ud { unlockData[id] = ud }
+			}
+			bleServerTransfer.unlockData = unlockData
 		}
+		try await self.bleServerTransfer.performDeviceEngagement(secureArea: SecureAreaRegistry.shared.get(name: secureAreaName), crv: crv)
+		return self.bleServerTransfer.status == .qrEngagementReady ? self.bleServerTransfer.qrCodePayload! : ""
 	}
 	
 	///  Receive request via BLE
@@ -58,17 +62,20 @@ public final class BlePresentationService: @unchecked Sendable, PresentationServ
 		}
 	}
 	
+	public func unlockKey(id: String) async throws -> Data? {
+		if let devicePrivateKey = bleServerTransfer.devicePrivateKeys[id] {
+			return try await devicePrivateKey.secureArea.unlockKey(id: id)
+		}
+		return nil
+	}
 	/// Send response via BLE
 	/// 
 	/// - Parameters:
 	///   - userAccepted: True if user accepted to send the response
 	///   - itemsToSend: The selected items to send organized in document types and namespaces
 	public func sendResponse(userAccepted: Bool, itemsToSend: RequestItems, onSuccess: ( @Sendable (URL?) -> Void)?) async throws  {
-		return try await withCheckedThrowingContinuation { c in
-			continuationResponse = c
-			handleSelected?(userAccepted, itemsToSend)
-			handleSelected = nil
-		}
+		await handleSelected?(userAccepted, itemsToSend)
+		handleSelected = nil
 	}
 }
 
@@ -82,28 +89,21 @@ extension BlePresentationService: MdocOfflineDelegate {
 				case .qrEngagementReady:
 						if let qrCode = self.bleServerTransfer.qrCodePayload {
 							deviceEngagement = qrCode
-							continuationQrCode?.resume(returning: qrCode)
-							continuationQrCode = nil
 						}
-				case .responseSent:
-					continuationResponse?.resume(returning: ())
-					continuationResponse = nil
 		default: break
 				}
 	}
 	/// Transfer finished with error
 	/// - Parameter error: The error description
 	public func didFinishedWithError(_ error: Error) {
-		continuationQrCode?.resume(throwing: error); continuationQrCode = nil
 		continuationRequest?.resume(throwing: error); continuationRequest = nil
-		continuationResponse?.resume(throwing: error); continuationResponse = nil
 	}
 	
 	/// Received request handler
 	/// - Parameters:
 	///   - request: Request items keyed by §UserRequestKeys§
 	///   - handleSelected: Callback function to call after user selection of items to send
-	public func didReceiveRequest(_ request: UserRequestInfo, handleSelected: @escaping (Bool, MdocDataTransfer18013.RequestItems?) -> Void) {
+	public func didReceiveRequest(_ request: UserRequestInfo, handleSelected: @escaping (Bool, MdocDataTransfer18013.RequestItems?) async -> Void) {
 		self.handleSelected = handleSelected
 		self.request = request
 		continuationRequest?.resume(returning: request)